+++ cmd/vegatools/commands/root.go
@@ -1,9 +1,6 @@
 package cmd
 
 import (
-	"fmt"
-	"os"
-
 	"github.com/spf13/cobra"
 )
 
@@ -15,11 +12,8 @@ var rootCmd = &cobra.Command{
 
 // Execute is the main function of `cmd` package.
 // Usually called by the `main.main()`.
-func Execute() {
-	if err := rootCmd.Execute(); err != nil {
-		fmt.Println(err)
-		os.Exit(1)
-	}
+func Execute() error {
+	return rootCmd.Execute()
 }
 
 func init() {