+++ cmd/vegatools/main.go
@@ -1,7 +1,0 @@
-package main
-
-import cmd "code.vegaprotocol.io/vega/cmd/vegatools/commands"
-
-func main() {
-	cmd.Execute()
-}